Add OpInfo tests that error out if sizes/strides is called in a composite context
This commit adds OpInfo tests that error out if sizes/strides get called
in a "composite context". A "composite context" is defined as a callsite
that is NOT in the backend implementation of an operator.

Everything that happens in a composite context gets traced by LazyTensor
or functorch, which is why we care about what happens here. If a
.sizes() or .strides() get called in a composite context, then today
these would get baked into the traced graph.